requirements.txt
django==4.1.4
djangorestframework==3.14.0
environs==9.5.0
Pillow==9.3.0